spotify-playlist-curator/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
numpy==1.26.0
google-generativeai==0.3.0
python-dotenv==1.0.0
pydantic==2.5.0
